New video for conference

+++ graphics.py
@@ -25,10 +25,10 @@ class Graphics:
                 img = pygame.image.load(image_path)
             print('in function...', image_path)
             if image_path == os.path.join('images', 'room_5'):
-                os.system('killall omxplayer.bin')
-                self.video_playing = False
+                if self.video_playing:
+                    os.system('killall omxplayer.bin')
+                    self.video_playing = False
                 omxc = Popen(['omxplayer', '-o',  'local', '--loop', self.path_2])
-                self.video_playing = True
                 img = None
             elif image_path == os.path.join('images', 'room_0', 'wall_0', '0.png'):
                 if not self.video_playing:
